🐞 fix: Suggestion panel disappears after mouse movement
The issue originated from the jQuery "caretMove" event patch, which failed to record the initial caret position when starting to listen for events. This caused the `caretMove` check to fail because the initial `caretPosition` was not updated, leading to false triggers whenever there was any keyboard or mouse activity.

Surprisingly, this issue went unnoticed for months. To address it, I’ve also isolated the `caretMove` checks for each new event listener to prevent similar issues in the future.